refactor: moves use cases to use-cases directory
