+++ src/use-cases/CalculateMealTicket.spec.ts
