+++ src/use-cases/CalculateMinSalary.spec.ts
