+++ src/use-cases/GetWorkCoefficient.spec.ts
