+++ src/use-cases/GetWorkCoefficient.ts
